This makes the uploaded dataset try to be imported in the external pages file.
