--- pages/8_🛑_Do_Not_Visit_Below_Pages.py
@@ -1,4 +1,0 @@
-import streamlit as st 
-
-st.title("Attention 👮🏿‍♀️🚨🚒👨🏻‍🚒")
-st.write("The below pages are meant to enable functionality. They are not meant to be visited.")
